+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # happy-mapper 1.0
 
+## happy-mapper-1.0.3-RELEASE
+
+* Fixes some rare special cases.
+
 ## happy-mapper-1.0.2-RELEASE
 
 * Adds sources and JavaDocs to build